% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr-rendering-single}
\alias{covr-rendering-single}
\alias{renderSourceRow}
\alias{renderSourceFile}
\alias{single_file_summary}
\title{Rendering for single file report}
\usage{
renderSourceRow(line, source, coverage)

renderSourceFile(lines, file = "source", highlight = TRUE)

single_file_summary(file_stats)
}
\arguments{
\item{line, lines}{Line(s) number}

\item{source}{source file}

\item{coverage}{The number of times covered}

\item{file}{the file in question}

\item{highlight}{Highlight the row.}

\item{file_stats}{The coverage object for the file.}
}
\description{
These functions facilitate the creation of reports for coverage of a
single file.
}
\details{
Coverage with test extraction
}
\concept{coverage}
