% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr-single}
\alias{covr-single}
\alias{file_coverage}
\alias{covr_file}
\title{Single File Coverage}
\usage{
file_coverage(file = rstudioapi::getSourceEditorContext()$path,
  pkg = ".", ...)

covr_file(x = file_coverage(), report.file = NULL,
  show.report = interactive())
}
\description{
These functions extract tests, run tests and create a report of the coverage for
a single file.
}
\section{Functions}{
\itemize{
\item \code{file_coverage}: Extract tests and compute the coverage for the given file.

\item \code{covr_file}: Create a report for a single
}}

