% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr_files}
\alias{covr_files}
\title{Compute coverage for a group of files.}
\usage{
covr_files(filter, pkg = ".", report = TRUE)
}
\arguments{
\item{filter}{An optional filter to restrict the files to extract from and run tests for.}

\item{pkg}{The package to test.}

\item{create}{Display a report of the coverage?}
}
\description{
Compute coverage for a group of files.
}
