% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_namespace.R
\name{is_namespace_registered}
\alias{is_namespace_registered}
\title{Check if a namespace is registered}
\usage{
is_namespace_registered(ns)
}
\arguments{
\item{ns}{a namespace environment or a character name of a namespace.}
}
\description{
Check if a namespace is registered
}
