% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-testing.R
\name{is_valid_regex}
\alias{is_valid_regex}
\title{Check if a regular expression is valid.}
\usage{
is_valid_regex(pattern)
}
\arguments{
\item{pattern}{the regular expression pattern to test.}
}
\description{
Check if a regular expression is valid.
}
