% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_namespace.R
\name{new_namespace_env}
\alias{new_namespace_env}
\alias{new_pkg_environment}
\title{Create a new namespace environment}
\usage{
new_namespace_env(name, path = file.path(tempdir()),
  import = "methods")

new_pkg_environment(name = "test package environment", ...,
  register = FALSE)
}
\arguments{
\item{name}{The name of the environment}

\item{path}{An optional path.}

\item{import}{Package to include in the imports.}

\item{...}{Arguments passed on to \code{new_namespace_env}
\describe{
  \item{name}{The name of the environment}
  \item{path}{An optional path.}
  \item{import}{Package to include in the imports.}
}}

\item{register}{Should the package namespace be registered?}
}
\description{
Create a new namespace environment
}
