% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_namespace.R
\name{register_namespace}
\alias{register_namespace}
\title{Register a namespace}
\usage{
register_namespace(ns)
}
\arguments{
\item{ns}{A namespace environment.}
}
\description{
Only full packages should ever be registered.
}
