% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{validity-tests}
\alias{validity-tests}
\alias{is_valid}
\alias{are_valid}
\title{Alternate check for validity}
\usage{
is_valid(object, complete = FALSE)

are_valid(lst, complete = FALSE)
}
\arguments{
\item{object}{an S4 object to test for validity}

\item{complete}{logical; if \code{TRUE}, \code{validObject} is
      called recursively for each of the slots.  The default is \code{FALSE}.}

\item{lst}{a list of S4 objects to test for validity.}
}
\description{
These functions will test if an object is valid
returning a value appropriate to use in \code{\link[assertthat:validate_that]{assertthat::validate_that()}},
\code{\link[assertthat:assert_that]{assertthat::assert_that()}}, or \code{\link[assertthat:see_if]{assertthat::see_if()}}.
}
\section{Functions}{
\itemize{
\item \code{is_valid}: Check if an object is valid.

\item \code{are_valid}: Check if each element in a list is valid.
}}

\seealso{
Other validity-tests: 
\code{\link{expect_valid}()}
}
\concept{validity-tests}
