% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_condition.R
\name{catch_condition}
\alias{catch_condition}
\title{Catch a condition for testing.}
\usage{
catch_condition(code)
}
\arguments{
\item{code}{code to run that should assert a condition.}
}
\description{
This function captures a condition object such as a warning or
error, to allow for testing components and classes.
}
