% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr_files}
\alias{covr_files}
\title{Compute coverage for a group of files.}
\usage{
covr_files(filter, pkg = ".", report = TRUE)
}
\arguments{
\item{filter}{A regular expression filter to apply to the files from \code{pkg}.}

\item{pkg}{The package to compute coverage for.}

\item{report}{If a report should be constructed and shown.}
}
\description{
Compute coverage for a group of files.
}
