% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{expect_valid}
\alias{expect_valid}
\title{Expect an S4 object is valid}
\usage{
expect_valid(object, complete = FALSE, label = NULL)
}
\arguments{
\item{object}{an S4 object to test for validity}

\item{complete}{logical; if \code{TRUE}, \code{validObject} is
      called recursively for each of the slots.  The default is \code{FALSE}.}

\item{label}{Used to customise failure messages. For expert use only.}
}
\description{
Similar to \code{\link[=is_valid]{is_valid()}} except designed to work in the
\code{\link[testthat:test_that]{testthat::test_that()}} framework.
}
\seealso{
Other validity-tests: 
\code{\link{validity-tests}}
}
\concept{validity-tests}
